remove reciperr and add spellcheck to precommit
