--- docs/files/builders/reciperr/overview.md
@@ -1,13 +1,0 @@
----
-hide:
-  - toc
----
-# Reciperr Builders
-
-You can find movies using a Reciperr list on [reciperr.com](https://reciperr.com/) (Reciperr). 
-
-No configuration is required for this Builder.
-
-| Builder                    | Description                                    |             Works with Movies              |             Works with Shows             |    Works with Playlists and Custom Sort    |
-|:---------------------------|:-----------------------------------------------|:------------------------------------------:|:----------------------------------------:|:------------------------------------------:|
-| [`reciperr_list`](list.md) | Finds every movie at a Reciperr JSON data URL. | :fontawesome-solid-circle-check:{ .green } | :fontawesome-solid-circle-xmark:{ .red } | :fontawesome-solid-circle-check:{ .green } |
